+++ Recommender.SDK.MVC/Recommender.SDK.MVC.csproj
@@ -1,18 +1,18 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <RootNamespace>KenticoCloud.Recommender.SDK</RootNamespace>
+    <TargetFrameworks>net461</TargetFrameworks>
+    <RootNamespace>KenticoCloud.Recommender.SDK.MVC</RootNamespace>
     <Description></Description>
     <TargetsForTfmSpecificBuildOutput>$(TargetsForTfmSpecificBuildOutput);IncludeP2PAssets</TargetsForTfmSpecificBuildOutput>
     <Copyright>(c) 2018 Kentico Software. All rights reserved.</Copyright>
-    <AssemblyTitle>Kentico Cloud Recommender AspNetCore SDK</AssemblyTitle>
-    <Title>Kentico Cloud Recommender AspNetCore SDK</Title>
+    <AssemblyTitle>Kentico Cloud Recommender SDK for .NET MVC</AssemblyTitle>
+    <Title>Kentico Cloud Recommender SDK for .NET MVC</Title>
     <NeutralLanguage>en-us</NeutralLanguage>
     <Authors>Kentico Software</Authors>
-    <AssemblyName>KenticoCloud.Recommender.SDK</AssemblyName>
-    <PackageId>KenticoCloud.Recommender.SDK</PackageId>
-    <PackageTags>kentico;mvc;aspnetcore;aspnetcoremvc</PackageTags>
+    <AssemblyName>KenticoCloud.Recommender.SDK.MVC</AssemblyName>
+    <PackageId>KenticoCloud.Recommender.SDK.MVC</PackageId>
+    <PackageTags>kentico;mvc</PackageTags>
     <PackageIconUrl>http://kentico.com/favicon.ico</PackageIconUrl>
     <PackageProjectUrl></PackageProjectUrl>
     <PackageLicenseUrl></PackageLicenseUrl>
@@ -29,13 +29,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Http.Extensions" Version="2.2.0" />
+    <ProjectReference Include="..\Recommender.SDK.Shared\Recommender.SDK.Shared.csproj" PrivateAssets="All"/>
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Recommender.SDK.Shared\Recommender.SDK.Shared.csproj" PrivateAssets="All"/>
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Web" />
   </ItemGroup>
 
   <Target Name="IncludeP2PAssets">
