+++ Recommender.SDK.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 15.0.28307.168
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Recommender.SDK", "Recommender.SDK\Recommender.SDK.csproj", "{3305E2D2-C398-4D56-9E3B-1BE299502FE3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Recommender.SDK.Shared", "Recommender.SDK.Shared\Recommender.SDK.Shared.csproj", "{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Recommender.SDK.MVC", "Recommender.SDK.MVC\Recommender.SDK.MVC.csproj", "{6B13C145-C46E-4A37-8B82-B2906B1FDE16}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{3305E2D2-C398-4D56-9E3B-1BE299502FE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3305E2D2-C398-4D56-9E3B-1BE299502FE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3305E2D2-C398-4D56-9E3B-1BE299502FE3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
