+++ Recommender.SDK/RecommendationClient.cs
@@ -1,42 +1,29 @@
 ﻿using System;
-using System.Text;
-using Newtonsoft.Json;
 using System.Net.Http;
-using System.Net.Http.Headers;
 using System.Threading.Tasks;
-using System.Web;
 using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
+using Recommender.SDK.Shared;
 
 namespace KenticoCloud.Recommender.SDK
 {
-    public partial class RecommendationClient : IRecommendationClient
+    public class RecommendationClient : RecommendationClientBase, IRecommendationClient
     {
-        private readonly HttpClient _httpClient;
-        private const string RecommendationApiRoutePrefix = "api/Recommend";
-        private readonly string _token;
-
         public RecommendationClient(string accessToken, int timeoutSeconds) : this("https://kc-recommender-api-beta.kenticocloud.com", accessToken, timeoutSeconds)
         {
         }
 
-        public RecommendationClient(string endpointUrl, string accessToken, int timeoutSeconds)
+        public RecommendationClient(string endpointUrl, string accessToken, int timeoutSeconds) : base(endpointUrl, accessToken, timeoutSeconds)
         {
-            _httpClient = new HttpClient { BaseAddress = new Uri(endpointUrl), Timeout = TimeSpan.FromSeconds(timeoutSeconds)};
-            _httpClient.DefaultRequestHeaders.Accept.Clear();
-            _httpClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-            _httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {accessToken}");
-            _token = accessToken;
         }
 
-        public RecommendationClient(HttpClient client, string accessToken)
+        public RecommendationClient(HttpClient client, string accessToken) : base(client, accessToken)
         {
-            _httpClient = client;
-            _token = accessToken;
         }
 
         private CallerInfo GetCallerInfoFromRequest(HttpRequest request, HttpResponse response, bool sessionBased)
         {
-            var projectId = TokenHelpers.GetProjectIdFromToken(_token);
+            var projectId = TokenHelpers.GetProjectIdFromToken(Token);
             if (string.IsNullOrEmpty(projectId))
                 throw new ArgumentException("Invalid authorization token.");
 
@@ -45,102 +32,6 @@ namespace KenticoCloud.Recommender.SDK
                 throw new ArgumentException("Uid has to be set.", nameof(cookie.Uid));
 
             return request.GetCallerInfo(cookie.Uid, cookie.Sid, sessionBased);
-        }
-
-        private async Task<T> GetAsync<T>(string url)
-        {
-            using (var response = await _httpClient.GetAsync(url))
-            {
-                var responseBody = await response.Content.ReadAsStringAsync();
-
-                if (response.IsSuccessStatusCode)
-                    return JsonConvert.DeserializeObject<T>(responseBody);
-
-                throw new RecommendationException(response.StatusCode, responseBody);
-            }
-        }
-
-        private async Task<T> PostAsync<T>(string url, string content)
-        {
-            using (var response = await _httpClient.PostAsync(url, new StringContent(content, Encoding.UTF8, "application/json")))
-            {
-                var responseBody = await response.Content.ReadAsStringAsync();
-
-                if (response.IsSuccessStatusCode)
-                    return JsonConvert.DeserializeObject<T>(responseBody);
-
-                throw new RecommendationException(response.StatusCode, responseBody);
-            }
-        }
-
-        private Task<RecommendedContentItem[]> GetRecommendationsFromRequest(RecommendationRequest request)
-        {
-            var queryString =
-                new StringBuilder($"{RecommendationApiRoutePrefix}/Items?currentItemId={request.Codename}&visitId={request.VisitId}&limit={request.Limit}");
-
-            if (!string.IsNullOrWhiteSpace(request.ContentTypeName) && request.ContentTypeName != "*")
-                queryString.Append($"&contentTypeName={request.ContentTypeName}");
-
-            if(!string.IsNullOrWhiteSpace(request.FilterQuery))
-                queryString.Append($"&filterQuery={request.FilterQuery}");
-
-            if(!string.IsNullOrWhiteSpace(request.BoosterQuery))
-                queryString.Append($"&boosterQuery={request.BoosterQuery}");
-
-            if(!string.IsNullOrWhiteSpace(request.SourceApp))
-                queryString.Append($"&sourceApp={request.SourceApp}");
-
-            if (request.SeparateTracking)
-                queryString.Append($"&separateTracking=true");
-
-            switch (request.Method)
-            {
-                case HttpMethodEnum.Get:
-                    return GetAsync<RecommendedContentItem[]>(queryString.ToString());
-                case HttpMethodEnum.Post:
-                    return PostAsync<RecommendedContentItem[]>(queryString.ToString(), JsonConvert.SerializeObject(request.VisitorData));
-                default:
-                    throw new Exception("Unspecified Htttp Method for Request");
-            }
-        }
-
-        public Task<RecommendedContentItem[]> GetRecommendationsAsync(
-            string visitId, 
-            string codename, 
-            int limit, 
-            string contentType,
-            string filterQuery = "", 
-            string boosterQuery = "", 
-            string sourceApp = "",
-            bool separateTracking = false)
-        {
-            if (string.IsNullOrEmpty(visitId))
-                throw new ArgumentException("Visit Id has to be set.", nameof(visitId));
-
-            if(string.IsNullOrEmpty(codename))
-                throw new ArgumentException("Codename has to be set.", nameof(codename));
-
-            if(string.IsNullOrEmpty(contentType))
-                throw new ArgumentException("Content Type has to be set.", nameof(contentType));
-
-            var projectId = TokenHelpers.GetProjectIdFromToken(_token);
-            if (string.IsNullOrEmpty(projectId))
-                throw new ArgumentException("Invalid authorization token.");
-
-            var request = new RecommendationRequest
-            {
-                Method = HttpMethodEnum.Get,
-                VisitId = visitId,
-                Codename = codename,
-                Limit = limit,
-                ContentTypeName = contentType,
-                FilterQuery = filterQuery,
-                BoosterQuery = boosterQuery,
-                SourceApp = sourceApp,
-                SeparateTracking = separateTracking
-            };
-
-            return GetRecommendationsFromRequest(request);
         }
 
         public Task<RecommendedContentItem[]> GetRecommendationsAsync(
@@ -178,6 +69,76 @@ namespace KenticoCloud.Recommender.SDK
             };
 
             return GetRecommendationsFromRequest(req);
+        }
+
+        public RecommendationRequest CreateRequest(HttpRequest request, HttpResponse response, string codename,
+            int limit, string contentType, bool sessionBased = false)
+        {
+            var visitorData = GetCallerInfoFromRequest(request, response, sessionBased);
+            var req = new RecommendationRequest
+            {
+                VisitId = visitorData.VisitId,
+                Codename = codename,
+                Limit = limit,
+                ContentTypeName = contentType,
+                VisitorData = visitorData,
+                Method = HttpMethodEnum.Post,
+                Executor = GetRecommendationsFromRequest
+            };
+
+            return req;
+        }
+
+         public Task TrackConversionAsync(
+            string codename, 
+            HttpRequest request, 
+            HttpResponse response, 
+            string sourceApp = "", 
+            bool sessionBased = false)
+        {
+            if (string.IsNullOrEmpty(codename))
+                throw new ArgumentException("Codename has to be set.", nameof(codename));
+
+            var callerInfo = GetCallerInfoFromRequest(request, response, sessionBased);
+            callerInfo.SourceApp = sourceApp;
+
+            return PostAsync<object>($"{TrackingApiRoutePrefix}/Conversion?visitId={callerInfo.VisitId}&contentItemId={codename}",
+                JsonConvert.SerializeObject(callerInfo));
+        }
+
+        public Task TrackPortionViewAsync(
+            string codename,
+            int portionPercentage,
+            HttpRequest request, 
+            HttpResponse response,
+            string sourceApp = "", 
+            bool sessionBased = false)
+        {
+            if (string.IsNullOrEmpty(codename))
+                throw new ArgumentException("Codename has to be set.", nameof(codename));
+
+            var callerInfo = GetCallerInfoFromRequest(request, response, sessionBased);
+            callerInfo.SourceApp = sourceApp;
+
+            return PostAsync<object>($"{TrackingApiRoutePrefix}/PortionView?visitId={callerInfo.VisitId}&contentItemId={codename}&portionPercentage={portionPercentage}",
+                JsonConvert.SerializeObject(callerInfo));
+        }
+
+        public Task TrackVisitAsync(
+            string codename, 
+            HttpRequest request, 
+            HttpResponse response, 
+            string sourceApp = "", 
+            bool sessionBased = false)
+        {
+            if (string.IsNullOrEmpty(codename))
+                throw new ArgumentException("Codename has to be set.", nameof(codename));
+
+            var callerInfo = GetCallerInfoFromRequest(request, response, sessionBased);
+            callerInfo.SourceApp = sourceApp;
+
+            return PostAsync<object>($"{TrackingApiRoutePrefix}/Visit?visitId={callerInfo.VisitId}&contentItemId={codename}",
+                JsonConvert.SerializeObject(callerInfo));
         }
     }
 }
